Deduplicate svs-like level image and metadata parsing
