+++ opentile/tiff_image.py
@@ -629,6 +629,59 @@ class NativeTiledTiffImage(BaseTiffImage, metaclass=ABCMeta):
         return frame_index
 
 
+class NativeTiledLevelImage(NativeTiledTiffImage, LevelTiffImage):
+    """A pyramid level whose tiles are natively tiled and served as-is, with scale,
+    pyramid index, and pixel spacing derived from the base level. Used by formats that
+    only need this plain passthrough level (e.g. Huron, Mikroscan, Motic)."""
+
+    def __init__(
+        self, page: TiffPage, file: OpenTileFile, base_size: Size, base_mpp: SizeMm
+    ):
+        """Parameters
+        ----------
+        page: TiffPage
+            TiffPage defining the level.
+        file: OpenTileFile
+            File to read data from.
+        base_size: Size
+            Size of the base level in the pyramid.
+        base_mpp: SizeMm
+            Pixel spacing (um/pixel) of the base level in the pyramid.
+        """
+        super().__init__(page, file)
+        self._base_size = base_size
+        self._base_mpp = base_mpp
+        self._scale = self._calculate_scale(base_size)
+        self._pyramid_index = self._calculate_pyramidal_index(self._scale)
+        self._mpp = self._calculate_mpp(base_mpp, self._scale)
+
+    def __repr__(self) -> str:
+        return (
+            f"{type(self).__name__}({self._page}, {self._file}, "
+            f"{self._base_size}, {self._base_mpp})"
+        )
+
+    @property
+    def supported_compressions(self) -> Optional[list[COMPRESSION]]:
+        return None
+
+    @property
+    def mpp(self) -> SizeMm:
+        return self._mpp
+
+    @property
+    def pixel_spacing(self) -> SizeMm:
+        return self.mpp / 1000
+
+    @property
+    def scale(self) -> float:
+        return self._scale
+
+    @property
+    def pyramid_index(self) -> int:
+        return self._pyramid_index
+
+
 class DecodedTiledTiffImage(BaseTiffImage, metaclass=ABCMeta):
     """Meta class for images that are not natively tiled and have no per-tile encoded
     representation (e.g. strip-stored or uncompressed pages). The whole page is decoded
